+++ src/options/options.html
@@ -1,31 +1,10 @@
 <!-- (c) 2022 Clyde D'Souza -->
 <html>
-    <head>
-        <meta charset="utf-8">
-        <title>Chrome Extension Template</title>
-        <meta name="viewport" content="initial-scale=1, maximum-scale=1" />
-        <meta name="description" content="Chrome Extension Template" />
-        <script src="options.js" type="text/javascript"></script>
-        <link href="options.css" rel="stylesheet" />
-    </head>
-    <body>
-        <main>
-            <form>
-                <label>
-                    Chrome storage value
-                    <input type="text" id="storageValue" aria-label="Chrome storage value" />
-                </label>
-                <br/>
-                <button id="SaveConfiguration" type="button" class="btn-primary">Save new Chrome Storage value</button>
-            </form>
-        </main>
-        <footer>
-            <a href="https://github.com/ClydeDz/google-meet-exit-page-chrome-extension/issues/new/choose" target="_blank" rel="noreferrer">
-                Visit this on GitHub</a><span class="link-separator"></span>
-            <a href="https://sponsor.clydedsouza.net/" target="_blank" rel="noreferrer">
-                Buy me a coffee</a><span class="link-separator"></span>
-            <a href="https://clydedsouza.net" target="_blank" rel="noreferrer">
-                Developed by Clyde D'Souza</a>
-        </footer>        
-    </body>
+  <head>
+    <meta charset="utf-8" />
+    <title>Option Pages</title>
+    <meta name="viewport" content="initial-scale=1, maximum-scale=1" />
+    <script src="options.js" type="text/javascript"></script>
+  </head>
+  <body></body>
 </html>
