+++ src/popup/popup.html
@@ -1,30 +1,28 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
+  <head>
+    <meta charset="UTF-8" />
     <title>Auto String</title>
     <script src="popup.js" type="text/javascript"></script>
-    <link href="popup.css" rel="stylesheet" />
-</head>
-<body>
+  </head>
+  <body>
     <div id="container">
-        <h2>Auto string filler</h2>
-        <div class="btn-row">
-            <div>
-                <label>Primary Color:</label>
-                <input type="color" id="primaryColor" value="#f1beb0">
-            </div>
-            <div>
-                <label>Background Color:</label>
-                <input type="color" id="backgroundColor" value="#424242">
-            </div>
-            <div>
-                <label>Border Color:</label>
-                <input type="color" id="borderColor" value="#323232">
-            </div>
+      <h2>Auto string filler</h2>
+      <div class="btn-row">
+        <div>
+          <label>Primary Color:</label>
+          <input type="color" id="primaryColor" value="#f1beb0" />
         </div>
-        <button id="save" class="default-btn">Save</button>
+        <div>
+          <label>Background Color:</label>
+          <input type="color" id="backgroundColor" value="#424242" />
+        </div>
+        <div>
+          <label>Border Color:</label>
+          <input type="color" id="borderColor" value="#323232" />
+        </div>
+      </div>
+      <button id="save" class="default-btn">Save</button>
     </div>
-</body>
+  </body>
 </html>
-
